api/backend/requirements.txt
# Django Backend Requirements
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0

# Google Gemini API (v1)
google-generativeai==0.3.2

# Export Libraries
reportlab==4.0.7
python-docx==1.1.0
Pillow==10.1.0

# GitHub API
PyGithub==1.59.1

# Development
python-dotenv==1.0.0